+++ client/my-sites/email/paths.js
@@ -78,12 +78,8 @@ export function emailManagementNewTitanAccount( siteName, domainName, relativeTo
 	return emailManagementEdit( siteName, domainName, 'titan/new', relativeTo );
 }
 
-export function emailManagementCompleteTitanMailboxSetup(
-	siteName,
-	domainName,
-	relativeTo = null
-) {
-	return emailManagementEdit( siteName, domainName, 'titan/complete-mailbox-setup', relativeTo );
+export function emailManagementTitanSetupMailbox( siteName, domainName, relativeTo = null ) {
+	return emailManagementEdit( siteName, domainName, 'titan/setup-mailbox', relativeTo );
 }
 
 export function emailManagementTitanControlPanelRedirect(
